+++ storybook-webgpu/src/atmosphere/CruisingAltitude-Story.tsx
@@ -6,9 +6,9 @@ import {
   type ThreeElement
 } from '@react-three/fiber'
 import { TilesPlugin } from '3d-tiles-renderer/r3f'
-import { Suspense, useRef, useState, type FC } from 'react'
+import { Suspense, useLayoutEffect, useRef, useState, type FC } from 'react'
 import { AgXToneMapping, Vector3, type Object3D } from 'three'
-import { mrt, output, pass, toneMapping, uniform } from 'three/tsl'
+import { context, mrt, output, pass, toneMapping, uniform } from 'three/tsl'
 import {
   MeshLambertNodeMaterial,
   PostProcessing,
@@ -22,7 +22,7 @@ import {
 } from '@takram/three-atmosphere'
 import {
   aerialPerspective,
-  AtmosphereContextNode,
+  AtmosphereContext,
   AtmosphereLight,
   AtmosphereLightNode,
   skyEnvironment
@@ -83,8 +83,15 @@ const Content: FC<StoryProps> = () => {
   const scene = useThree(({ scene }) => scene)
   const camera = useThree(({ camera }) => camera)
 
-  const atmosphereContext = useResource(() => new AtmosphereContextNode(), [])
+  const atmosphereContext = useResource(() => new AtmosphereContext(), [])
   atmosphereContext.camera = camera
+
+  useLayoutEffect(() => {
+    renderer.contextNode = context({
+      ...renderer.contextNode.value,
+      getAtmosphere: () => atmosphereContext
+    })
+  }, [renderer, atmosphereContext])
 
   // Post-processing:
 
@@ -104,8 +111,8 @@ const Content: FC<StoryProps> = () => {
   const velocityNode = passNode.getTextureNode('velocity')
 
   const aerialNode = useResource(
-    () => aerialPerspective(atmosphereContext, colorNode, depthNode),
-    [atmosphereContext, colorNode, depthNode]
+    () => aerialPerspective(colorNode, depthNode),
+    [colorNode, depthNode]
   )
 
   const lensFlareNode = useResource(() => lensFlare(aerialNode), [aerialNode])
@@ -215,16 +222,13 @@ const Content: FC<StoryProps> = () => {
     )
   })
 
-  const envNode = useResource(
-    () => skyEnvironment(atmosphereContext),
-    [atmosphereContext]
-  )
+  const envNode = useResource(() => skyEnvironment(), [])
   scene.environmentNode = envNode
 
   return (
     <>
       <atmosphereLight
-        args={[atmosphereContext, 40]}
+        args={[40]}
         castShadow
         shadow-normalBias={0.1}
         shadow-mapSize={[2048, 2048]}