+++ storybook-webgpu/src/atmosphere/Vanilla-WorldOriginRebasing.tsx
@@ -8,7 +8,7 @@ import {
   Vector3
 } from 'three'
 import { OrbitControls } from 'three/addons/controls/OrbitControls.js'
-import { pass, toneMapping } from 'three/tsl'
+import { context, pass, toneMapping } from 'three/tsl'
 import {
   MeshPhysicalNodeMaterial,
   PostProcessing,
@@ -21,7 +21,7 @@ import {
   getSunDirectionECI
 } from '@takram/three-atmosphere'
 import {
-  AtmosphereContextNode,
+  AtmosphereContext,
   AtmosphereLight,
   AtmosphereLightNode,
   skyBackground
@@ -62,18 +62,22 @@ async function init(container: HTMLDivElement): Promise<() => void> {
 
   // The atmosphere context manages resources like LUTs and uniforms shared by
   // multiple nodes:
-  const context = new AtmosphereContextNode()
+  const atmosphereContext = new AtmosphereContext()
+  renderer.contextNode = context({
+    ...renderer.contextNode.value,
+    getAtmosphere: () => atmosphereContext
+  })
 
   // Create a scene with a sky background:
   const scene = new Scene()
-  scene.backgroundNode = skyBackground(context).add(dithering)
+  scene.backgroundNode = skyBackground().add(dithering)
 
   // Move and rotate the ellipsoid so that the world origin locates at
   // the ECEF coordinates, and the scene's orientation aligns with
   // x: north, y: up, z: east.
   Ellipsoid.WGS84.getNorthUpEastFrame(
     positionECEF,
-    context.matrixWorldToECEF.value
+    atmosphereContext.matrixWorldToECEF.value
   )
 
   // Create a torus knot inside the group:
@@ -93,7 +97,7 @@ async function init(container: HTMLDivElement): Promise<() => void> {
   // Create the atmospheric light. Note that this story omits the atmospheric
   // scattering, which is only plausible when the distance between the camera
   // and scene objects is small enough to ignore it.
-  const light = new AtmosphereLight(context)
+  const light = new AtmosphereLight()
   light.castShadow = true
   light.distance = 1 // Distance from the light target to the light
   light.shadow.camera.top = 1
@@ -126,23 +130,23 @@ async function init(container: HTMLDivElement): Promise<() => void> {
     camera.updateMatrixWorld()
     observerECEF
       .setFromMatrixPosition(camera.matrixWorld)
-      .applyMatrix4(context.matrixWorldToECEF.value)
+      .applyMatrix4(atmosphereContext.matrixWorldToECEF.value)
 
     // Configure the planetary conditions in the atmosphere context according to
     // the current date and optionally the point of observation:
     const currentDate = +date + ((clock.getElapsedTime() * 5e6) % 864e5)
     const matrixECIToECEF = getECIToECEFRotationMatrix(
       currentDate,
-      context.matrixECIToECEF.value
+      atmosphereContext.matrixECIToECEF.value
     )
     getSunDirectionECI(
       currentDate,
-      context.sunDirectionECEF.value,
+      atmosphereContext.sunDirectionECEF.value,
       observerECEF
     ).applyMatrix4(matrixECIToECEF)
     getMoonDirectionECI(
       currentDate,
-      context.moonDirectionECEF.value,
+      atmosphereContext.moonDirectionECEF.value,
       observerECEF
     ).applyMatrix4(matrixECIToECEF)
 
@@ -165,7 +169,7 @@ async function init(container: HTMLDivElement): Promise<() => void> {
     controls.dispose()
     geometry.dispose()
     material.dispose()
-    context.dispose()
+    atmosphereContext.dispose()
     renderer.dispose()
   }
 }