+++ types/three.d.ts
@@ -1,5 +1,12 @@
 import type { Camera } from 'three'
-import type { Node, NodeFrame, Renderer, UniformNode } from 'three/webgpu'
+import type {
+  ContextNode,
+  Node,
+  NodeBuilderContext,
+  NodeFrame,
+  Renderer,
+  UniformNode
+} from 'three/webgpu'
 
 import type { NodeType, NodeValueTypeOf } from '@takram/three-geospatial/webgpu'
 
@@ -25,9 +32,13 @@ declare module 'three/tsl' {
 }
 
 declare module 'three/webgpu' {
-  // Add "camera"
+  interface Renderer {
+    contextNode: ContextNode
+  }
+
   interface NodeBuilder {
     camera?: Camera
+    context: NodeBuilderContext
   }
 
   interface Node {